Request Jellyfin's termination - don't kill it
JF has no main window for CloseMainWindow() to send WM_CLOSE to, so it will always get killed. This becomes apparent if you look in C:\ProgramData\Jellyfin\Server\data and see the extra files for the database with non-committed changes.

This works by attaching to the console and sending Ctrl+C. A little hackish.
A psuedoconsole might work better but the tray app wouldn't be able to control an existing process.

The best approach is probably sending a request via Jellyfin's API to shutdown.